please iwyu and clang

+++ src/activity_actor_definitions.h
@@ -2443,6 +2443,7 @@ class pulp_activity_actor : public activity_actor
         // either single tripoint shoved into set, or 3x3 zone around u/npc
         std::set<tripoint_abs_ms> placement;
 
+        // NOLINT(cata-serialize)
         float float_corpse_damage_accum = 0.0f;
 
         int unpulped_corpses_qty = 0;
@@ -2455,7 +2456,7 @@ class pulp_activity_actor : public activity_actor
         // if corpse cost more than hour to pulp, drop it
         bool way_too_long_to_pulp = false;
 
-        double bash_factor;
+        float bash_factor;
         int cut_quality;
         // bools for end message
         bool stomps_only = false;
@@ -2475,6 +2476,7 @@ class pulp_activity_actor : public activity_actor
         // how far the splatter goes
         int mess_radius;
         // what `placement` we are currently at
+        // NOLINT(cata-serialize)
         std::set<tripoint_abs_ms>::const_iterator current_pos_iter;
 };
